déplacement de l'algo de recherche dans une page html recherche, dans un controller creneauRDV
