+++ templates/base_template.html.twig
@@ -28,7 +28,7 @@
             <div class="col">
                 <ul class="navbar-nav row w-100 justify-content-center text-center">
                     <li class="col-md-auto nav-item d-flex justify-content-center align-items-center fs-4 me-3">
-                        <a class="nav-link" {% if menu=='recherche' %} active {% endif %} aria-current="page" href="index.php?controleur=agenda&methode=obtenir">Rechercher un créneau</a>
+                        <a class="nav-link" {% if menu=='recherche' %} active {% endif %} aria-current="page" href="index.php?controleur=creneauRDV&methode=obtenir">Rechercher un créneau</a>
                     </li>
                     <li class="col-md-auto nav-item d-flex justify-content-center align-items-center fs-4 me-3">
                         <a class="nav-link" {% if menu=='agenda' %} active {% endif %} aria-current="page" href="index.php?controleur=agenda&methode=lister">Agendas</a>
